Moving Ethernet type to network_types
